Split upscale library assets hook
